# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CSI St. Andrews Sithalapakkam API",
        "description": "API documentation for CSI St. Andrews Sithalapakkam",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/csistandrewssithalapakkam_api/public",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/verses/today": {
            "get": {
                "tags": [
                    "Verses"
                ],
                "summary": "Get verse for today",
                "description": "Get verse for today/current week",
                "operationId": "510e742530b3a185b744cfdecaec742a",
                "responses": {
                    "200": {
                        "description": "Today's verse",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "verse_id": {
                                            "type": "integer"
                                        },
                                        "week_date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "verse_tamil": {
                                            "type": "string"
                                        },
                                        "verse_english": {
                                            "type": "string"
                                        },
                                        "verse_status": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No verse found"
                    }
                }
            }
        },
...
        "/api/verses/{id}": {
            "get": {
                "tags": [
                    "Verses"
                ],
                "summary": "Get verse by ID",
                "description": "Get verse by ID",
                "operationId": "d49075d9debecb345ab14547b555875c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verse details"
                    },
                    "404": {
                        "description": "Verse not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based based auth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}